move internal client codegen to pkg level
