+++ client/pkg/codegen/utils/utils.go
