+++ client/pkg/codegen/utils/utils_test.go
